pyproject.toml
[project]
name = "nethackrc"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "argparse>=1.4.0",
    "jinja2>=3.1.6",
    "pyquery>=2.0.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
